refactor(api-headless-cms-ddb-es): reuse es createSort method
